Fix strandedness issue in bug #14
This path fixes the issue reported in github issue [#14](https://github.com/harrispopgen/mutyper/issues/14).
The issue originates from bisect.bisect using 1-based indexes, as
shown [here](https://docs.python.org/3.9/library/bisect.html), causing
the script to not extract the index correctly when the variant fall out
or within the last index. Also, it causes to intersect with the wrong
interval in the next if statement, since pointing to the index+1
position in the list.